style: 42 norm

+++ philo_bonus/src/main.c
@@ -6,7 +6,7 @@
 /*   By: troberts <troberts@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/17 18:16:35 by troberts          #+#    #+#             */
-/*   Updated: 2023/02/05 23:10:56 by troberts         ###   ########.fr       */
+/*   Updated: 2023/02/06 00:34:29 by troberts         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,10 +15,8 @@
 void	create_semaphors(int number_philosophers, sem_t **forks, sem_t **output)
 {
 	sem_unlink(SEM_NAME_FORK);
-	*forks = sem_open(SEM_NAME_FORK, O_CREAT, S_IRUSR | S_IWUSR, number_philosophers);
-	int sem_value;
-	sem_getvalue(*forks, &sem_value); // Remove
-	printf("%i\n%i\n",number_philosophers, sem_value);
+	*forks = sem_open(SEM_NAME_FORK, O_CREAT, S_IRUSR | S_IWUSR,
+			number_philosophers);
 	if (*forks == SEM_FAILED)
 		exit(EXIT_FAILURE);
 	sem_unlink(SEM_NAME_OUTPUT);
@@ -68,36 +66,40 @@ t_philo	create_philo_struct(int ac, char **av)
 	return (philo);
 }
 
+void	wait_child(t_philo philo_struct, pid_t *philos_pid)
+{
+	int			wstatus;
+	int			i;
+
+	i = 0;
+	while (i < philo_struct.common.nbr_philosophers)
+	{
+		waitpid(-1, &wstatus, 0);
+		if (WIFEXITED(wstatus) && WEXITSTATUS(wstatus) == DEAD_PHILO)
+		{
+			i = 0;
+			while (i < philo_struct.common.nbr_philosophers)
+			{
+				kill(philos_pid[i], SIGKILL);
+				i++;
+			}
+			return ;
+		}
+		i++;
+	}
+	return ;
+}
+
 int	main(int ac, char **av)
 {
 	pid_t		*philos_pid;
 	t_philo		philo_struct;
-	int			wstatus;
-	int			i;
 
 	if (ac < 5 || ac > 6)
 		return (EXIT_FAILURE);
 	philo_struct = create_philo_struct(ac, av);
 	philos_pid = launch_philos(philo_struct);
-	i = 0;
-	while (i < philo_struct.common.nbr_philosophers)
-	{
-		waitpid(-1, &wstatus, 0);
-		if (WIFEXITED(wstatus))
-		{
-			if (WEXITSTATUS(wstatus) == DEAD_PHILO)
-			{
-				i = 0;
-				while (i < philo_struct.common.nbr_philosophers)
-				{
-					kill(philos_pid[i], SIGKILL);
-					i++;
-				}
-				break ;
-			}
-		}
-		i++;
-	}
+	wait_child(philo_struct, philos_pid);
 	clean_philos(philo_struct);
 	return (EXIT_SUCCESS);
 }
